requirements.txt
numpy==1.26.4
pillow==10.2.0
torch==2.2.2
torchvision==0.17.2
scikit-learn==1.4.1.post1
matplotlib==3.8.3
streamlit==1.31.1
pickle-mixin==1.0.2
